% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circleSector.R
\name{addCircleSectorMid}
\alias{addCircleSectorMid}
\title{Add Circle sectors to a Leaflet map from a dataframe using bearing and spread}
\usage{
addCircleSectorMid(
  map,
  data,
  lat = "lat",
  lon = "lon",
  bearing = "bearing",
  angle = "angle",
  radius = "radius",
  weight = "weight",
  colour = "colour",
  label = "label",
  popup = "popup"
)
}
\arguments{
\item{map}{A leaflet map}

\item{data}{A data frame where each row represents on circle sector and
columns for lat, lon, min, mid, max.}

\item{lat}{The column name as string holding the Latitude of the circle mid
point on the map in WGS 84 decimal degrees. Default: "lat".}

\item{lon}{The column name as string holding the Longitude of the circle mid
point on the map in WGS 84 decimal degrees. Default: "lon".}

\item{bearing}{The column name as string holding the main bearing of the sector
in decimal degrees, counted clockwise from North. Default: "bearing".}

\item{angle}{max The column name as string holding the angle of the sector
in decimal degrees, which will be centered on the main direction \code{mid}.
Default: "angle".}

\item{radius}{The column name as string holding the radius of the circle in
metres. Default: "radius".}

\item{weight}{The column name as string holding the weight of the circle.
Default: "weight.}

\item{colour}{The column name as string holding the colour of the circle.
Default: "colour". Example values: "blue", "red", "#3388ff".}

\item{label}{The column name as string holding the label content of the circle.
Default: "label".}

\item{popup}{The column name as string holding the popup content of the circle.
Default: "popup".}
}
\description{
This methor wraps the \code{setSector} method of
\href{https://github.com/kluizeberg/Leaflet.Circle-sector}{Leaflet.Circle-sector},
a Leaflet plugin extending \href{https://leafletjs.com/index.html#circle}{L.Circle}.
}
\examples{
\dontrun{
data("fans_mean")

leaflet_basemap(l_height = 500, l_width = 700) \%>\%
  leaflet::clearBounds() \%>\%
  addCircleSectorMid(data = fans_mean)
}
}
\seealso{
Other plugin: 
\code{\link{addCircleSectorMinMax}()},
\code{\link{circlePlugin}()},
\code{\link{leaflet_basemap}()},
\code{\link{registerPlugin}()}
}
\concept{plugin}
