% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circlePlugin.R
\name{circlePlugin}
\alias{circlePlugin}
\title{Register htmlDependency for leaflet.circle-sector.js}
\usage{
circlePlugin()
}
\description{
This method invokes \code{htmltools::htmlDependency} with preset arguments to
the dependency \code{Leaflet.Circle-sector}.
}
\seealso{
Other plugin: 
\code{\link{addCircleSectorMid}()},
\code{\link{addCircleSectorMinMax}()},
\code{\link{leaflet_basemap}()},
\code{\link{registerPlugin}()}
}
\concept{plugin}
