% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fans}
\alias{fans}
\title{Turtle hatchling nest emergences: Main fan of tracks}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 129 rows and 69 columns.
}
\usage{
fans
}
\description{
Nest fans are described by
\itemize{
\item encounter_source_id A UUID for the TurtleNestEncounter
\item encounter_when Datetime of the encounter
\item encounter_latitude Latitude of the encounter in decimal degrees WGS84
\item encounter_laongitude Longitude of the encounter in decimal degrees WGS84
\item encounter_name The nest tag if known, a unique ID for the nest
\item encounter_species The species of turtle which laid that nest
\item bearing_leftmost_track_degrees The leftmost fan angle bearing
\item bearing_rightmost_track_degrees The rightmost fan angle bearing
\item no_tracks_main_group The number of hatchling tracks in the main group of
tracks, best estimate
\item no_tracks_main_group_min The number of hatchling tracks in the main group of
tracks, lowest estimate
\item no_tracks_main_group_max The number of hatchling tracks in the main group of
tracks, highest estimate
\item outlier_tracks_present Whether outlier tracks were present
\item path_to_sea_comments Any comments on the path to sea, e.g. whether obstructed
\item hatchling_emergence_time_known Whether the exact time of the emergence is known
\item light_sources_present
\item hatchling_emergence_time
\item hatchling_emergence_time_accuracy
\item cloud_cover_at_emergence
\item datetime
\item calendar_date_awst
\item calendar_date_awst_text
\item calendar_year
\item turtle_date
\item turtle_date_awst_text
\item season
\item season_week
\item iso_week
}
}
\seealso{
Other included: 
\code{\link{fans_mean}},
\code{\link{fans_tracks}},
\code{\link{fans_water}},
\code{\link{light_segments_artificial}},
\code{\link{light_segments_natural}},
\code{\link{lights}},
\code{\link{outlier_segments}},
\code{\link{outliers}}
}
\concept{included}
\keyword{datasets}
