% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fans_water}
\alias{fans_water}
\title{A data frame ready to be used in \code{addCircleSectorMid}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 129 rows and 9 columns.
}
\usage{
fans_water
}
\description{
#' This data frame is created from \code{fans} in \code{inst/generate_data.R}.
}
\details{
Columns:
\itemize{
\item lat
\item lon
\item mid
\item angle
\item radius
\item weight
\item colour
\item popup
}
}
\seealso{
Other included: 
\code{\link{fans_mean}},
\code{\link{fans_tracks}},
\code{\link{fans}},
\code{\link{light_segments_artificial}},
\code{\link{light_segments_natural}},
\code{\link{lights}},
\code{\link{outlier_segments}},
\code{\link{outliers}}
}
\concept{included}
\keyword{datasets}
