% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lights}
\alias{lights}
\title{Turtle hatchling nest emergences: Known light sources}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 553 rows and 59 columns.
}
\usage{
lights
}
\description{
With the same base columns as fans:
\itemize{
\item bearing_light_degrees
\item light_source_type
\item light_source_description
}
}
\seealso{
Other included: 
\code{\link{fans_mean}},
\code{\link{fans_tracks}},
\code{\link{fans_water}},
\code{\link{fans}},
\code{\link{light_segments_artificial}},
\code{\link{light_segments_natural}},
\code{\link{outlier_segments}},
\code{\link{outliers}}
}
\concept{included}
\keyword{datasets}
