% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{outliers}
\alias{outliers}
\title{Turtle hatchling nest emergences: Individual outlier tracks}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 187 rows and 59 columns.
}
\usage{
outliers
}
\description{
With the same base columns as in dataset fans:
\itemize{
\item bearing_outlier_track_degrees The compass bearing of the outlier track 5m
from the nest in degrees
\item outlier_group_size The number of tracks in the outlier group, typically 1
\item outlier_track_comment Any further comments on the outlier track
}
}
\seealso{
Other included: 
\code{\link{fans_mean}},
\code{\link{fans_tracks}},
\code{\link{fans_water}},
\code{\link{fans}},
\code{\link{light_segments_artificial}},
\code{\link{light_segments_natural}},
\code{\link{lights}},
\code{\link{outlier_segments}}
}
\concept{included}
\keyword{datasets}
