% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registerPlugin.R
\name{registerPlugin}
\alias{registerPlugin}
\title{A pipe-friendly way to register a plugin for a Leaflet map}
\usage{
registerPlugin(map, plugin)
}
\arguments{
\item{map}{A Leaflet map, eg. the output of \code{leaflet::leaflet()}}

\item{plugin}{A plugin, eg. \code{leaflet.circlesector::circlePlugin}}
}
\value{
The map with the plugin registered
}
\description{
A pipe-friendly way to register a plugin for a Leaflet map
}
\examples{
\dontrun{
library(leaflet.circlesector)  # loads pipe
leaflet::leaflet() \%>\%
   registerPlugin(circlePlugin())
}
}
\seealso{
Other plugin: 
\code{\link{addCircleSectorMid}()},
\code{\link{addCircleSectorMinMax}()},
\code{\link{circlePlugin}()},
\code{\link{leaflet_basemap}()}
}
\concept{plugin}
